fix: use merge commits
For detecting change we rely on merge commits instead of reading all commits
from the history. This broke before our changelog, as we merged
git history when moving into a mono-repo.

As other tools didn't used merged commits before, we use this
as solution to detect only changes done in the zeebe repo.